requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coverage==5.2.1
coveralls==2.1.2
defusedxml==0.6.0
Django==3.1
django-allauth==0.42.0
django-cors-headers==3.5.0
django-filter==2.3.0
django-oauth-toolkit==1.3.2
django-rest-auth==0.9.5
django-simple-history==2.11.0
djangorestframework==3.11.1
docopt==0.6.2
flake8==3.8.3
gunicorn==20.0.4
idna==2.10
Markdown==3.2.2
mccabe==0.6.1
oauthlib==3.1.0
Pillow==7.2.0
pycodestyle==2.6.0
pyflakes==2.2.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
